refactor: separate concerns for telemetry validation
