+++ internal/controller/telemetryvalidation_controller.go
@@ -2,13 +2,8 @@ package controller
 
 import (
 	"context"
-	"crypto/rand"
 	"fmt"
-	"maps"
-	"slices"
-	"strings"
 
-	"k8s.io/apimachinery/pkg/api/meta"
 	"k8s.io/apimachinery/pkg/runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
@@ -16,7 +11,6 @@ import (
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	ctrl "sigs.k8s.io/controller-runtime"
 	logger "sigs.k8s.io/controller-runtime/pkg/log"
 
@@ -28,32 +22,6 @@ type TelemetryValidationReconciler struct {
 
 	Scheme *runtime.Scheme
 }
-
-const (
-	correlationProcessorName                       = "attributes/correlation_id"
-	correlationResourceProcessorName               = "resource/correlation_id"
-	correlationDDTagsProcessorName                 = "transform/correlation_ddtags"
-	correlationMetricsProcessorName                = "transform/metrics_correlation_id"
-	correlationAttributeKey                        = "correlation_id"
-	correlationHeaderFromCtxKey                    = "metadata.x-correlation-id"
-	correlationDDTagKey                            = "correlation_id:"
-	setDDTagsOnlyStatement                         = `set(attributes["ddtags"], Concat(["%s", attributes["%s"]], "")) where attributes["%s"] != nil and attributes["ddtags"] == nil`
-	appendDDTagsStatement                          = `set(attributes["ddtags"], Concat([attributes["ddtags"], ",", "%s", attributes["%s"]], "")) where attributes["%s"] != nil and attributes["ddtags"] != nil`
-	setMetricCorrelationStatement                  = `set(attributes["%s"], resource.attributes["%s"]) where attributes["%s"] == nil and resource.attributes["%s"] != nil`
-	deleteMetricDDTagsStatement                    = `delete_key(attributes, "ddtags") where attributes["ddtags"] != nil`
-	deleteMetricTagsStatement                      = `delete_key(attributes, "tags") where attributes["tags"] != nil`
-	defaultValidatorImage                          = "ghcr.io/mydecisive/mdai-fidelity-validator:0.1.0"
-	defaultValidatorPort                     int32 = 18081
-	defaultValidatorReceiverPort             int32 = 8126
-	defaultValidatorReplicas                 int32 = 1
-	telemetryValidationLabelKey                    = "hub.mydecisive.ai/telemetry-validation"
-	telemetryValidationRunIDAnnotationKey          = "hub.mydecisive.ai/telemetry-validation-run-id"
-	telemetryValidationRunIDMetricLabel            = "telemetry_validation_run_id"
-	telemetryValidationPrometheusSourceLabel       = "__meta_kubernetes_service_label_hub_mydecisive_ai_telemetry_validation"
-	telemetryValidationRunIDPrometheusSource       = "__meta_kubernetes_service_annotation_hub_mydecisive_ai_telemetry_validation_run_id"
-	telemetryValidationRoleShadow                  = "telemetry-validation-shadow-collector"
-	telemetryValidationRoleValidator               = "telemetry-validation-validator"
-)
 
 // +kubebuilder:rbac:groups=hub.mydecisive.ai,resources=telemetryvalidations,verbs=get;list;watch;create;update;patch;delete
 // +kubebuilder:rbac:groups=hub.mydecisive.ai,resources=telemetryvalidations/status,verbs=get;update;patch
@@ -111,78 +79,4 @@ func (r *TelemetryValidationReconciler) SetupWithManager(mgr ctrl.Manager) error
 		Owns(&corev1.Service{}).
 		Owns(&appsv1.Deployment{}).
 		Complete(r)
-}
-
-func activeSignals(signals []hubv1.TelemetrySignal) []hubv1.TelemetrySignal {
-	if len(signals) == 0 {
-		return []hubv1.TelemetrySignal{
-			hubv1.TelemetrySignalMetrics,
-			hubv1.TelemetrySignalLogs,
-			hubv1.TelemetrySignalTraces,
-		}
-	}
-
-	unique := make([]hubv1.TelemetrySignal, 0, len(signals))
-	for _, signal := range signals {
-		if !slices.Contains(unique, signal) {
-			unique = append(unique, signal)
-		}
-	}
-
-	return unique
-}
-
-func setValidationCondition(conditions *[]metav1.Condition, generation int64, status metav1.ConditionStatus, reason, message string) {
-	meta.SetStatusCondition(conditions, metav1.Condition{
-		Type:               typeAvailableHub,
-		Status:             status,
-		Reason:             reason,
-		Message:            message,
-		ObservedGeneration: generation,
-	})
-}
-
-func mergeMaps(a, b map[string]string) map[string]string {
-	if a == nil && b == nil {
-		return nil
-	}
-
-	out := make(map[string]string, len(a)+len(b))
-	maps.Copy(out, a)
-	maps.Copy(out, b)
-	return out
-}
-
-func resolveTelemetryValidationRunID(validation *hubv1.TelemetryValidation) (string, bool, error) {
-	specified := strings.TrimSpace(validation.Spec.RunID)
-	if specified != "" {
-		return specified, specified != validation.Status.RunID, nil
-	}
-	if strings.TrimSpace(validation.Status.RunID) != "" {
-		return validation.Status.RunID, false, nil
-	}
-
-	generated, err := generateTelemetryValidationRunID()
-	if err != nil {
-		return "", false, err
-	}
-	return generated, true, nil
-}
-
-func generateTelemetryValidationRunID() (string, error) {
-	var b [16]byte
-	if _, err := rand.Read(b[:]); err != nil {
-		return "", fmt.Errorf("generate telemetry validation run id: %w", err)
-	}
-
-	b[6] = (b[6] & 0x0f) | 0x40 //nolint:mnd
-	b[8] = (b[8] & 0x3f) | 0x80 //nolint:mnd
-	return fmt.Sprintf(
-		"%08x-%04x-%04x-%04x-%012x",
-		b[0:4],
-		b[4:6],
-		b[6:8],
-		b[8:10],
-		b[10:16],
-	), nil
 }
